+++ graylog2-web-interface/src/theme/colors.js
@@ -1,5 +1,58 @@
+// @flow strict
 import teinte from './variants/teinte';
-// TODO: Create more themes
+
+export type Colors = {
+  brand: {
+    primary: string,
+    secondary: string,
+    tertiary: string,
+  },
+  global: {
+    textDefault: string,
+    textAlt: string,
+    background: string,
+    contentBackground: string,
+    link: string,
+    linkHover: string,
+  },
+  gray: {
+    "0": string,
+    "10": string,
+    "20": string,
+    "30": string,
+    "40": string,
+    "50": string,
+    "60": string,
+    "70": string,
+    "80": string,
+    "90": string,
+    "100": string,
+  },
+  variant: {
+    danger: string,
+    default: string,
+    info: string,
+    primary: string,
+    success: string,
+    warning: string,
+    light: {
+      danger: string,
+      default: string,
+      info: string,
+      primary: string,
+      success: string,
+      warning: string,
+    },
+    dark: {
+      danger: string,
+      default: string,
+      info: string,
+      primary: string,
+      success: string,
+      warning: string,
+    },
+  },
+};
 
 const colors = {
   teinte,
