+++ graylog2-web-interface/src/theme/types.js
@@ -1,0 +1,6 @@
+// @flow strict
+import { type Colors } from './colors';
+
+export type ThemeInterface = {
+  colors: Colors,
+};
